src/requirements.txt
anyio==4.3.0
certifi==2024.2.2
dependency-injector==4.41.0
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.4
httpx==0.26.0
idna==3.6
python-telegram-bot==20.8
six==1.16.0
sniffio==1.3.1
typing_extensions==4.10.0
